Fix deadlock in RegionizedTaskQueue reference counting
The ticket removal logic is inside of the compute() call on
reference count map. The compute() call will acquire a lock on
the table bin entry. The compute() call will then possibly attempt
a ticket removal - while holding the table bin entry lock.

Usual usage of the ticket removal logic does not hold the ticket
lock for the area being removed. However, scheduling of the
ChunkFullTask may be done during ticket level update (i.e
cancelling the task). Thus, it is possible to enter the compute
call holding the ticket lock. This allows for a deadlock to occur
as the lock acquire order is not always consistent.

To fix this, we move the ticket removal outside of the compute()
call. This does require changing the ticket to include an ID
to avoid race conditions where a ticket is removed while a new
one has already been added.