Update Paper
This update includes a new scheduler which may be used.
Currently, there are two schedulers that may be configured
in the global config file:
- EDF
- WORK_STEALING

The EDF is simply the old scheduler has used since inception.
The new WORK_STEALING scheduler provides possibly better scheduling:
- intermediate task execution
  i.e chunk loading and packet processing may happen inbetween ticks,
  lowering latency for these tasks
- NUMA aware scheduling on Linux with the `-DPaper.NumaScheduling=true`
  startup flag
- better thread/NUMA locality, as the scheduler attempts to keep regions
  on the same scheduler thread/node
- dynamic thread count adjustment

The EDF scheduler is default as it has proven to be reliable,
even if lacking.